Fix path optimization check (#254)
I think this check was wrong. It should only skip the optimization when exclusionArea is used not when it is not used. This might also be missing exclusionAreaPlace and exclusionAreaBreak but I am 90% sure that that dose not matter anyway as the optimization step cannot introduce any new blocks to break or place.